Fix env vars and missing steps
Neither GOOGLE_APPLICATION_CREDENTIALS nor GCP_SA_FILE appear to be
environment variables that are referenced anywhere in this repository,
and do not work at runtime. Replace both with TF_VAR_gcp_sa_file_loc,
which sets the variable the Terraform is looking for.

Correct the name of the osd binary to "ocm".

Clarify that ocm and gcloud must be logged in before running.