Add Workload Identity Configuration shell resource
Add new gcp_authentication_type variable, which defaults to
"service_account" for backward compatability, but allows for specifying
"workload_identity_federation" to use WIF.

When using WIF, a new shell script resource controls running the ocm
commands to create or delete the wif-config. This includes a "read"
action which does a "describe" and then manually converts to JSON, since
that is what the shell provider needs.

Note that the resource is not actually used to alter the behavior of
cluster creation yet.